+++ R/01_linear_labels.R
@@ -22,6 +22,5 @@ linear_labels <- S7::new_generic("linear_labels", "granule")
 # #' @rdname linear_labels
 method(linear_labels, mt_unit) <- function(granule, i, label = FALSE, abbreviate = TRUE, ...) {
   if (!label) return(as.character(i))
-  # TODO - pluralise the full time unit
-  paste0(i, if (abbreviate) time_unit_abbr(granule) else time_unit_full(granule))
+  paste0(i, if (abbreviate) time_unit_abbr(granule) else time_unit_plural(granule, i))
 }